% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfactor.R
\name{as.bitfactor}
\alias{as.bitfactor}
\title{Convert a standard factor to a bitfactor}
\usage{
as.bitfactor(x)
}
\arguments{
\item{x}{factor object}
}
\value{
bitfactor object
}
\description{
Convert a standard factor to a bitfactor
}
