% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytefactor.R
\name{as.bytefactor}
\alias{as.bytefactor}
\title{Convert an ordinary factor into a bytefactor}
\usage{
as.bytefactor(x)
}
\arguments{
\item{x}{factor object}
}
\value{
bytefactor object if number of levels < 256, otherwise error
}
\description{
Convert an ordinary factor into a bytefactor
}
