% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytefactor.R
\name{as.character.bytefactor}
\alias{as.character.bytefactor}
\alias{as.integer.bytefactor}
\alias{as.double.bytefactor}
\title{Other representations of a small factor}
\usage{
\method{as.character}{bytefactor}(x, ...)

\method{as.integer}{bytefactor}(x, ...)

\method{as.double}{bytefactor}(x, ...)
}
\arguments{
\item{x}{bytefactor object}

\item{...}{other arguments passed to \code{as.character.factor()}}
}
\description{
Other representations of a small factor
}
