% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfactor.R
\name{as.character.smallfactor}
\alias{as.character.smallfactor}
\title{Character representation of a small factor}
\usage{
\method{as.character}{smallfactor}(x, ...)
}
\arguments{
\item{x}{smallfactor object}

\item{...}{other arguments passed to \code{as.character.factor()}}
}
\description{
Character representation of a small factor
}
