% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytefactor.R
\name{as.data.frame.bytefactor}
\alias{as.data.frame.bytefactor}
\title{Include a bytefactor in a data.frame object}
\usage{
\method{as.data.frame}{bytefactor}(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  nm = deparse1(substitute(x))
)
}
\arguments{
\item{x, row.names, optional, ..., nm}{see documentation for \code{as.data.frame.faactor()}}
}
\value{
data.frame
}
\description{
Code based on \code{as.data.frame.numeric}
}
