% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfactor.R
\name{as.data.frame.smallfactor}
\alias{as.data.frame.smallfactor}
\title{Include a smallfactor in a data.frame object}
\usage{
\method{as.data.frame}{smallfactor}(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  nm = deparse1(substitute(x))
)
}
\arguments{
\item{x, row.names, optional, ..., nm}{see documentation for \code{as.data.frame.faactor()}}
}
\value{
data.frame
}
\description{
Include a smallfactor in a data.frame object
}
