% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfactor.R
\name{as.factor.bitfactor}
\alias{as.factor.bitfactor}
\title{Convert a bitfactor to a factor}
\usage{
as.factor.bitfactor(x, ...)
}
\arguments{
\item{x}{bitfactor}

\item{...}{ignored}
}
\value{
standard R factor
}
\description{
Convert a bitfactor to a factor
}
