% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytefactor.R
\name{as.factor.bytefactor}
\alias{as.factor.bytefactor}
\title{Convert a bytefactor into a regular factor}
\usage{
as.factor.bytefactor(x)
}
\arguments{
\item{x}{bytefactor object}
}
\description{
Convert a bytefactor into a regular factor
}
