% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfactor.R
\name{as.factor.smallfactor}
\alias{as.factor.smallfactor}
\title{Convert a smallfactor into a regular factor}
\usage{
as.factor.smallfactor(x)
}
\arguments{
\item{x}{smallfactor object}
}
\description{
Convert a smallfactor into a regular factor
}
