% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfactor.R
\name{as.integer.bitfactor}
\alias{as.integer.bitfactor}
\alias{as.double.bitfactor}
\alias{as.character.bitfactor}
\alias{length.bitfactor}
\title{convert bitfactor to integer, numeric and character}
\usage{
\method{as.integer}{bitfactor}(x, ...)

\method{as.double}{bitfactor}(x, ...)

\method{as.character}{bitfactor}(x, ...)

\method{length}{bitfactor}(x, ...)
}
\arguments{
\item{x}{bitfactor}

\item{...}{ignored}
}
\description{
convert bitfactor to integer, numeric and character
}
