% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfactor.R
\name{as.smallfactor}
\alias{as.smallfactor}
\title{Convert an ordinary factor into a smallfactor}
\usage{
as.smallfactor(x)
}
\arguments{
\item{x}{factor object}
}
\value{
smallfactor object if number of levels < 256, otherwise error
}
\description{
Convert an ordinary factor into a smallfactor
}
