% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfactor.R
\name{bitfactor}
\alias{bitfactor}
\title{Create a bitfactor}
\usage{
bitfactor(x, levels = NULL)
}
\arguments{
\item{x}{something coercile into a character vector}

\item{levels}{named levels.  This should probably just come from the
values on 'x'.}
}
\value{
bitfactor object
}
\description{
Create a bitfactor
}
