% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytefactor.R
\name{bytefactor}
\alias{bytefactor}
\title{Create a bytefactor i.e. a factor backed by a raw byte vector, not an integer}
\usage{
bytefactor(x, levels = NULL)
}
\arguments{
\item{x}{object to turn into a factor}

\item{levels}{the levels of the factor. Default: NULL means to take the
level names from the given object}
}
\description{
Create a bytefactor i.e. a factor backed by a raw byte vector, not an integer
}
