% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{chunk}
\alias{chunk}
\title{Split a vector in chunks of size 'n'}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{chunk size}
}
\value{
list of vectors
}
\description{
Split a vector in chunks of size 'n'
}
