% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_into_single_int}
\alias{pack_into_single_int}
\title{Pack integer values into a single integer}
\usage{
pack_into_single_int(small_ints, nbits)
}
\arguments{
\item{small_ints}{vector of small integer values. these are recycle if necessary}

\item{nbits}{number of bits we want to pack these into}
}
\value{
single integer value
}
\description{
Pack integer values into a single integer
}
