% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytefactor.R
\name{print.bytefactor}
\alias{print.bytefactor}
\title{Print a bytefactor}
\usage{
\method{print}{bytefactor}(x, ...)
}
\arguments{
\item{x}{bytefactor object}

\item{...}{other arguments passed to \code{print.factor()}}
}
\description{
Print a bytefactor
}
