% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfactor.R
\name{print.smallfactor}
\alias{print.smallfactor}
\title{Print a smallfactor}
\usage{
\method{print}{smallfactor}(x, ...)
}
\arguments{
\item{x}{smallfactor object}

\item{...}{other arguments passed to \code{print.factor()}}
}
\description{
Print a smallfactor
}
