% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfactor.R
\name{[.bitfactor}
\alias{[.bitfactor}
\title{Subset a bitfactor}
\usage{
\method{[}{bitfactor}(x, y)
}
\arguments{
\item{x}{bitfactor}

\item{y}{subset spedification}
}
\description{
ToDo: This currently converts the entire bitfactor into a regular R
factor, subsets that R factor, then converts the subset back into a
bitfactor.   This is horribly inefficient, and should really just extract
the values being asked for.
}
