% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfactor.R
\name{[.smallfactor}
\alias{[.smallfactor}
\title{Subset a smallfactor}
\usage{
\method{[}{smallfactor}(x, ...)
}
\arguments{
\item{x}{smallfactor}

\item{...}{ignored}
}
\description{
Subset a smallfactor
}
