% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{unpack_from_single_int}
\alias{unpack_from_single_int}
\title{Unpacka an integer value into a vector of small integers of 'nbits' each}
\usage{
unpack_from_single_int(int, nbits)
}
\arguments{
\item{int}{single integer value}

\item{nbits}{number of bits we want to pack these into}
}
\value{
vector of small integers
}
\description{
Unpacka an integer value into a vector of small integers of 'nbits' each
}
