% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{unpack_ints}
\alias{unpack_ints}
\title{Unpack a 32-bit integer values into a vector of small integers of 'nbits' each}
\usage{
unpack_ints(ints, nbits)
}
\arguments{
\item{ints}{integer vector}

\item{nbits}{number of bits we want to pack these into}
}
\value{
vector of small integers
}
\description{
Unpack a 32-bit integer values into a vector of small integers of 'nbits' each
}
